refac(portforward): code organization http logging
