+++ crates/kftray-commons/src/utils/mod.rs
@@ -3,6 +3,5 @@ pub mod config_dir;
 pub mod config_state;
 pub mod db;
 pub mod github;
-pub mod logging;
 pub mod migration;
 pub mod validate_configs;
