+++ crates/kftray-portforward/src/lib.rs
@@ -1,17 +1,24 @@
-pub mod client;
-pub mod core;
-pub mod models;
-pub mod pod_finder;
+pub mod http_logs;
+pub mod kube;
 pub mod port_forward;
 
-mod tcp_forwarder;
-mod udp_forwarder;
-
-pub use core::*;
-
-pub use client::*;
-pub use models::*;
-pub use pod_finder::*;
-pub use port_forward::*;
-pub use tcp_forwarder::*;
-pub use udp_forwarder::*;
+pub use http_logs::{
+    HttpLogState,
+    Logger,
+};
+pub use kube::client::*;
+pub use kube::models::{
+    KubeContextInfo,
+    KubeNamespaceInfo,
+    KubeServiceInfo,
+    KubeServicePortInfo,
+    PodInfo,
+};
+pub use kube::{
+    deploy_and_forward_pod,
+    retrieve_service_configs,
+    start_port_forward,
+    stop_all_port_forward,
+    stop_port_forward,
+    stop_proxy_forward,
+};
