+++ crates/kftui/src/tui/input/navigation.rs
@@ -1,7 +1,7 @@
 use kftray_commons::models::config_model::Config;
 use kftray_commons::utils::config::insert_config;
-use kftray_portforward::client::list_kube_contexts;
-use kftray_portforward::core::retrieve_service_configs;
+use kftray_portforward::kube::client::list_kube_contexts;
+use kftray_portforward::kube::retrieve_service_configs;
 
 use crate::core::port_forward::{
     start_port_forwarding,
